Title: 사각형 채우기 3, Time: 89ms, Memory: 9MB, Status: Passed - Codetree

+++ 260209/사각형 채우기 3/rectangle-fill-3.java
@@ -8,7 +8,7 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
 		int n = Integer.parseInt(br.readLine());
-		int[] arr = new int[1001];
+		long[] arr = new long[1001];
 
 		arr[0] = 1;
 		arr[1] = 2;
